requirements.txt
groq
fastapi
uvicorn
streamlit
requests
python-dotenv
beautifulsoup4
ddgs
